Refactor resolution handling and polling for YouTube and Twitch
